pyproject.toml
[project]
authors = [
    {name = "Dmitry Kolomatskiy", email = "58207913+lemontree210@users.noreply.github.com"},
]
license = {text = "GPL-3.0-only"}
requires-python = "<3.11, >=3.10"
dependencies = [
    "python-telegram-bot[webhooks]>=21.4,<23.0",
    "phonenumberslite>=8.13.42,<10.0.0",
    "python-dotenv<2.0.0,>=1.0.1",
    "uvicorn<1.0.0,>=0.30.5",
    "starlette<1.0.0,>=0.38.2",
]
name = "samanthas-telegram-bot"
version = "0.1.0"
description = "Telegram bot(s) for Samanthas Group"

[dependency-groups]
dev = [
    "black<25.0.0,>=24.4.2",
    "ruff<1.0.0,>=0.5.5",
    "pre-commit<4.0.0,>=3.8.0",
    "icecream<3.0.0,>=2.1.3",
    "isort<6.0.0,>=5.13.2",
    "mypy<2.0.0,>=1.11.1",
    "pytest<9.0.0,>=8.3.2",
]

[tool.black]
line-length = 99
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 99

[tool.mypy]
ignore_missing_imports = true
python_version = "3.10"
strict = true
implicit_optional = true
disallow_untyped_calls = false
warn_return_any = false

[[tool.mypy.overrides]]
module = [
  "tests.*"
]
ignore_errors = true

[tool.ruff]
line-length = 99

# Assume Python 3.10.
target-version = "py310"

[tool.pdm.dev-dependencies]
dev = [
    "black<25.0.0,>=24.4.2",
    "ruff<1.0.0,>=0.5.5",
    "pre-commit<4.0.0,>=3.8.0",
    "icecream<3.0.0,>=2.1.3",
    "isort<6.0.0,>=5.13.2",
    "mypy<2.0.0,>=1.11.1",
    "pytest<9.0.0,>=8.3.2",
]

[tool.pdm.build]
includes = [
    "samanthas_telegram_bot",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


